Final Fixes - destroying image with listing , validating Price

+++ controllers/listings.js
@@ -28,11 +28,18 @@ module.exports.showListing = async (req,res) => {
 };
 
 module.exports.createListing = async (req,res,next) => {
+
+    // Enforce image required on create
+    if (!req.file) {
+      req.flash("error", "Image is required");
+      return res.redirect("/listings/new");
+    }
     let url = req.file.secure_url ;
     let filename = req.file.public_id ;
     const newListing = new Listing(req.body.listing) ;
     newListing.owner = req.user._id ;
     newListing.image = {url,filename} ;
+
     await newListing.save() ;
     req.flash("success","New listing created") ;
     res.redirect("/listings") ;
@@ -55,17 +62,13 @@ module.exports.updateListing = async (req, res) => {
   if (!req.body.listing) {
     throw new ExpressError(400, "Send Valid data for listing");
   }
-
   let { id } = req.params;
-
   // Fetch listing first
   let listing = await Listing.findById(id);
-
   if (!listing) {
     req.flash("error", "Listing not found");
     return res.redirect("/listings");
   }
-
   // Update basic fields manually
   listing.title = req.body.listing.title;
   listing.description = req.body.listing.description;
@@ -75,7 +78,6 @@ module.exports.updateListing = async (req, res) => {
 
   // If new image uploaded
   if (req.file) {
-
     // Only delete if old image exists
     if (listing.image && listing.image.filename) {
       await cloudinary.uploader.destroy(listing.image.filename);
@@ -86,17 +88,25 @@ module.exports.updateListing = async (req, res) => {
       filename: req.file.public_id
     };
   }
-
   await listing.save();
 
   req.flash("success", "Listing Updated Successfully");
   res.redirect(`/listings/${id}`);
 };
 
-module.exports.destroyListing = async (req,res) => {
-  let {id} = req.params ;
-  let deletedListing = await Listing.findByIdAndDelete(id);
-  console.log(deletedListing);
-  req.flash("success","Listing deleted") ;
-  res.redirect("/listings") ;
+module.exports.destroyListing = async (req, res) => {
+  let { id } = req.params;
+  let listing = await Listing.findById(id);
+
+  if (!listing) {
+    req.flash("error", "Listing not found");
+    return res.redirect("/listings");
+  }
+  // delete image from Cloudinary if exists
+  if (listing.image && listing.image.filename) {
+    await cloudinary.uploader.destroy(listing.image.filename);
+  }
+  await Listing.findByIdAndDelete(id);
+  req.flash("success", "Listing deleted");
+  res.redirect("/listings");
 };
